bot/requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==24.2.0
cairocffi==1.7.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cssselect2==0.7.0
defusedxml==0.7.1
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pillow==11.0.0
propcache==0.2.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple==3.8
requests==2.32.3
SQLAlchemy==2.0.36
tinycss2==1.4.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
webencodings==0.5.1
yarl==1.18.3
